consolidate handlers.go file
